refactor(senior): dedupe semLabel/yearLabel + neutral ledger-book audit entityType
Item 6 gate follow-ups (all three gates flagged, non-blocking):
- semLabel/yearLabel now have ONE definition in the pure Item 8 module
  (lib/score-ledger/stpshs-sheet.ts); the STPSHS data builder imports it and
  lib/score-ledger/ledger-book.ts re-exports it. Removes the 3-line clone that
  could drift and print divergent SHS period labels across the two artifacts.
- ledger-book route audit entityType 'senior_score_ledger' -> 'ledger_book'
  (the blank book never touches that table).

No behaviour change. typecheck/lint/build clean, 137 tests pass.
